mv the ilit to new folder, clear the output of ipy
