+++ AI-and-Analytics/Getting-Started-Samples/IntelLowPrecisionOptimization_GettingStarted/iLiT-Sample-for-Tensorflow/alexnet.yaml
@@ -1,0 +1,18 @@
+framework:
+  - name: tensorflow                              # possible values are tensorflow, mxnet and pytorch
+    inputs: x
+    outputs: Identity
+
+calibration:                                         
+  - iterations: 5, 10
+    algorithm:
+        activation: minmax
+
+tuning:
+    metric:  
+      - topk: 1
+    accuracy_criterion:
+      - relative: 0.01                              # the tuning target of accuracy loss percentage: 1%
+    timeout: 0                                      # tuning timeout (seconds)
+    random_seed: 9527                               # random seed                       
+
