+++ AI-and-Analytics/Getting-Started-Samples/IntelLowPrecisionOptimization_GettingStarted/iLiT-Sample-for-Tensorflow/run_jupyter.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+jupyter notebook --notebook-dir=./ --ip=0.0.0.0 --no-browser  --allow-root &
