--- Frameworks/iLiT/iLiT-Sample-for-Tensorflow/run_jupyter.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-jupyter notebook --notebook-dir=./ --ip=0.0.0.0 --no-browser  --allow-root &
